refactor: give owner resolution and window creation their own module
Finding which window owns a web view, and creating a fresh window to hold one,
are what both the routing and the move policy are built on: `findOwner` answers
five call sites of which only one is a move, and `createFreshWindow` serves the
open path and the move path alike. They move together into one module so the move
policy can import them without importing the router back.

`WebViewWindowCreator`, `WINDOW_CREATOR_WIRING_TIMEOUT_MS`, the creator state and
its wiring latch, `setWebViewWindowCreator`, `WindowShard`, `WebViewOwner`,
`findOwner`, `getLayoutTargetTabId`, `findLayoutTargetOwner`, `FreshWindow`,
`createFreshWindow` and `resetWindowCreatorForTesting` move verbatim; the only
edits are the `export` each needs and the imports each file no longer uses.
`openWebViewInOwningWindow` stays in the router: the move path never calls it.

`getWebViewShard` moves down to the shard index in the same commit rather than the
previous one, because this module turned out to call it and leaving it in the
router would have pointed this module back at the router. The router re-exports it,
so `main.ts` and `scroll-group-navigation.commands.ts` are untouched, as are
`setWebViewWindowCreator`'s six importers.
